fix: fixed too many rerenders issue in mentoring room Page before, I used custom hooks to listen to window resize by adding listener and add it to window resize event. even added debounce to negate too aggresive firings of the callback. but the mentor room page still is rendered up to 4 times! not good. so to fix this, I just settled by creating a variable WINDOW_HEIGHT and set it to a specific num, then use it on the parent page container, and just pass it down to the prop of the children who is dependent to the parent's height. if you can do it simply, why make it complicated? :)
