+++ src/hooks/useWindowSize.js
@@ -7,13 +7,15 @@ function debounce(fn, ms = 500) {
       timer = setTimeout(() => {
          timer = null
          fn.apply(this, arguments)
+        //  fn()
       }, ms)
    }
 }
 
 export default function useWindowSize() {
    const [size, setSize] = useState({width: 0, height: 0})
-
+    // console.log('use windows fired!')
+    
    useEffect(() => {
       const debouncedUpdateSizeHandler = debounce(function updateSize() {
          setSize({width: window.innerWidth, height: window.innerHeight})
