+++ src/index.css
@@ -110,17 +110,17 @@ label {
 
 .scrollbar::-webkit-scrollbar-track
 {
-	-webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
-	background-color: #26272bd1;
+	-webkit-box-shadow: inset 0 0 6px rgba(255, 255, 255, 0.3);
+	background-color: #ffffffd1;
 }
 
 .scrollbar::-webkit-scrollbar
 {
 	width: 6px;
-	background-color: #F5F5F5;
+	background-color: #ffffff;
 }
 
 .scrollbar::-webkit-scrollbar-thumb
 {
-	background-color: #303035;
+	background-color: #c8c8c8;
 }
