+++ src/pages/mentoring/room/MentoringRoomPage.jsx
@@ -1,38 +1,30 @@
-import PageContainer from '../../../components/PageContainer'
-import ParticipantsSideBar from './ParticipantsSideBar'
-import StreamActions from './StreamActions'
-import participantsMock from '../../../data/participantsMock.json'
-import chatsMock from '../../../data/chatsMock.json'
-import { useLayoutEffect, useRef, useState } from 'react'
-import ChatsSideBar from './ChatsSideBar'
-import useWindowSize from '../../../hooks/useWindowSize'
+import PageContainer from "../../../components/PageContainer"
+import ParticipantsSideBar from "./containers/ParticipantsSideBar"
+import StreamActions from "./containers/StreamActions"
+import participantsMock from "../../../data/participantsMock.json"
+import chatsMock from "../../../data/chatsMock.json"
+import ChatsSideBar from "./containers/ChatsSideBar"
 
 export default function MentoringRoomPage() {
-   const ref = useRef(null)
-   const {height} = useWindowSize()
-   const [pageHeight, setPageHeight] = useState(0)
+    const WINDOW_HEIGHT = 80
 
-   useLayoutEffect(() => {
-      setPageHeight(ref.current.offsetHeight)
-   }, [height])
+    return (
+        <div style={{ minHeight: `${WINDOW_HEIGHT}vh` }}>
+            <PageContainer
+                noPadding
+                className="min-h-[80vh] relative flex bg-gray-200"
+            >
+                <ParticipantsSideBar
+                    participantsCount={27}
+                    participants={participantsMock}
+                    pageHeight={WINDOW_HEIGHT}
+                />
+                <div className="flex-[1] relative z-[1]">
 
-   return (
-      <div ref={ref}>
-         <PageContainer noPadding className="min-h-[80vh] relative flex">
-            <ParticipantsSideBar
-               participantsCount={27}
-               participants={participantsMock}
-               pageHeight={pageHeight}
-            />
-            <div className='flex-[1] relative z-[1] bg-gray-800'>
-               <StreamActions />
-
-            </div>
-            <ChatsSideBar
-               chats={chatsMock}
-               pageHeight={pageHeight}
-            />
-         </PageContainer>
-      </div>
-   )
+                    <StreamActions />
+                </div>
+                <ChatsSideBar chats={chatsMock} pageHeight={WINDOW_HEIGHT} />
+            </PageContainer>
+        </div>
+    )
 }
